Refactor user list to user search endpoint and update related functions and tests
