--- openapi/paths/user/list/index.yaml
@@ -1,44 +1,0 @@
-parameters:
-  - name: nextCursor
-    in: query
-    description: nextCursor
-    schema:
-      type: string
-  - name: limit
-    in: query
-    description: limit items (default 10)
-    schema:
-      type: string
-  - name: sortOrder
-    in: query
-    description: asc or desc
-    schema:
-      type: string
-post:
-  tags:
-    - Users
-  operationId: usersList
-  summary: POST Get all users
-  description: This endpoint get all users
-  requestBody:
-    content:
-      application/json:
-        schema:
-          $ref: body.yaml
-  responses:
-    "200":
-      description: "Response"
-      content:
-        application/json:
-          schema:
-            $ref: "./response.yaml"
-    "400":
-      $ref: "../../../responses/bad.yaml"
-    "401":
-      $ref: "../../../responses/unauthorized.yaml"
-    "403":
-      $ref: "../../../responses/forbidden.yaml"
-    "404":
-      $ref: "../../../responses/not-found.yaml"
-
-  security: []
