--- openapi/paths/user/list/response.yaml
@@ -1,23 +1,0 @@
-type: object
-properties:
-  success:
-    type: boolean
-    example: true
-  payload:
-    type: object
-    properties:
-      nextCursor:
-        type: string
-      results:
-        type: array
-        items:
-          $ref: "../_types/user.yaml"
-      totalCount:
-        type: number
-    required:
-      - results
-      - total
-
-required:
-  - success
-  - payload
